fix: Import of moved classes
